Fix lint problems
