+++ src/worker.js
@@ -1,4 +1,3 @@
- 
 import { pipeline, WhisperTextStreamer } from "@huggingface/transformers";
 
 // Define model factories
